feat: bzlmodify
