+++ .bazelignore
@@ -1,1 +1,2 @@
+test
 compile_commands.json
