+++ BUILD.bazel
@@ -1,5 +1,4 @@
 load("@rules_cc//cc:defs.bzl", "cc_library")
-load("@rules_ecsact//ecsact:toolchain.bzl", "ecsact_toolchain")
 load("//bazel:copts.bzl", "copts")
 
 package(default_visibility = ["//visibility:public"])
@@ -13,36 +12,4 @@ cc_library(
     name = "lib",
     hdrs = glob(["ecsact/**/*.hh"]),
     copts = copts,
-)
-
-#### ECSACT TOOLCHAIN STUFF FOR TESTS BELOW HERE ####
-
-ecsact_toolchain(
-    name = "ecsact_cli_windows",
-    target_tool = "@ecsact_cli_windows//file",
-)
-
-ecsact_toolchain(
-    name = "ecsact_cli_linux",
-    target_tool = "@ecsact_cli_linux//file",
-)
-
-toolchain(
-    name = "ecsact_windows_toolchain",
-    exec_compatible_with = [
-        "@platforms//os:windows",
-        "@platforms//cpu:x86_64",
-    ],
-    toolchain = ":ecsact_cli_windows",
-    toolchain_type = "@rules_ecsact//ecsact:toolchain_type",
-)
-
-toolchain(
-    name = "ecsact_linux_toolchain",
-    exec_compatible_with = [
-        "@platforms//os:linux",
-        "@platforms//cpu:x86_64",
-    ],
-    toolchain = ":ecsact_cli_linux",
-    toolchain_type = "@rules_ecsact//ecsact:toolchain_type",
 )
