--- bazel/platforms/BUILD.bazel
@@ -1,19 +1,0 @@
-package(default_visibility = ["//:__subpackages__"])
-
-platform(
-    name = "windows",
-    constraint_values = [
-        "@platforms//cpu:x86_64",
-        "@platforms//os:windows",
-        "@bazel_tools//tools/cpp:msvc",
-    ],
-)
-
-platform(
-    name = "linux",
-    constraint_values = [
-        "@platforms//cpu:x86_64",
-        "@platforms//os:linux",
-        "@bazel_tools//tools/cpp:clang",
-    ],
-)
