+++ test/.bazelrc
@@ -1,0 +1,3 @@
+import %workspace%/../bazel/common.bazelrc
+
+try-import %workspace%/user.bazelrc
