+++ test/MODULE.bazel
@@ -1,0 +1,20 @@
+module(
+    name = "ecsact_rt_entt_test",
+    version = "0.1.0",
+    compatibility_level = 1,
+)
+
+bazel_dep(name = "rules_cc", version = "0.0.9")
+bazel_dep(name = "bazel_skylib", version = "1.4.2")
+bazel_dep(name = "googletest", version = "1.14.0")
+bazel_dep(name = "rules_ecsact", version = "0.4.0")
+bazel_dep(name = "ecsact_lang_cpp", version = "0.3.0")
+bazel_dep(name = "ecsact_runtime", version = "0.5.1")
+bazel_dep(name = "boost.mp11", version = "1.83.0.bzl.1")
+bazel_dep(name = "entt", version = "3.12.2")
+bazel_dep(name = "ecsact_rt_entt")
+
+local_path_override(
+    module_name = "ecsact_rt_entt",
+    path = "..",
+)
